+++ Scripts/Runtime/LaserScan/LaserScanPublisher.cs
@@ -68,7 +68,7 @@ public class LaserScanPublisher : MonoBehaviour
             // Update ROS Message
             uint sec = (uint)Math.Truncate(this._timeStamp);
             uint nanosec = (uint)( (this._timeStamp - sec)*1e+9 );
-            this._message.header.stamp.sec = sec;
+            this._message.header.stamp.sec = (int)sec;
             this._message.header.stamp.nanosec = nanosec;
             this._message.ranges = this._lidar.distances.ToArray();
             this._message.intensities = this._lidar.intensities.ToArray();
