+++ Scripts/Runtime/RGBCamera/RGBCameraPublisher.cs
@@ -48,7 +48,7 @@ public class RGBCameraPublisher : MonoBehaviour
             // Update ROS Message
             uint sec = (uint)Math.Truncate(this._timeStamp);
             uint nanosec = (uint)( (this._timeStamp - sec)*1e+9 );
-            this._message.header.stamp.sec = sec;
+            this._message.header.stamp.sec = (int)sec;
             this._message.header.stamp.nanosec = nanosec;
             this._message.data = this._camera.data;
             this._ros.Send(this._topicName, this._message);
